refactor(moderation): consolidate queries into queries.rs
