+++ server/src/moderation/mod.rs
@@ -8,9 +8,9 @@ pub mod error;
 pub mod filter_cache;
 pub mod filter_engine;
 pub mod filter_handlers;
-pub mod filter_queries;
 pub mod filter_types;
 pub mod handlers;
+pub mod queries;
 pub mod types;
 
 pub use error::ModerationError;
